First implementation for GithubActionsWorkflow Generator and Builder. With this Implementation three Builder are introduced and one Generator: The Builders are to be considered a general feature.
Builders:
* WorkflowBuilder.ts --> is able to generate a Workflow.
* JobBuilder.ts --> create a Job
* StepBuilder.ts --> creat a Step

Only the WorkflowBuilder class is needed to create a new File since it provides a fluent api and the other builders, it also is context aware of the current Job.

The Generator is specific to creating a GithubActions-File which from a Pipeline.ts which was itself generated from a GithubActions-File.