Implemented experimental conversion to StalkCD: jobs.<job_id>.steps[*].with https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepswith
args and entrypoint are omitted and got their own ParsingImpossibleReason-Enum-Type

If with is a simple string (jobs.<job_id>.steps[*].with = someValue) it is ignored and a ParsingImpossibleError is thrown with the reason ParsingImpossibleReason.StepWith

The Meta-Model was extended the IStep-Interface got an optional "environment" attribute. This allows adding environment-variables.